requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
python-telegram-bot==21.3
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
